python/packages/kagent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "kagent"
version = "0.3.0"
description = "kagent is a tool for building and deploying agent-based applications."
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
  "typer>=0.15.0",
  "protobuf>=6.31.1",
  "opentelemetry-api>=1.32.0",
  "opentelemetry-sdk>=1.32.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.32.0",
  "opentelemetry-instrumentation-openai>= 0.39.0",
  "opentelemetry-instrumentation-httpx >= 0.52.0",
  "fastapi>=0.115.1",
  "aiofiles>=24.1.0",
  "kagent-adk>=0.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kagent"]

[project.scripts]
kagent = "kagent.cli:run"